Fixed routing for super_admin candidate editing
p much same thing as local_admin fix but forgot to add description so will do it here. When clicking the edit button through the Candidates -> specific position then edit button route vs the allCandidates -> edit button route, it would direct you to the same editCandidate screen. So I made the routing all lead back to the view election. Since if you went down one path certain parameters would be passed which wouldnt be obtained through the other path. Probably a better fix instead of a generic route for the back button and the redirect after the update but this is what I came up w for now.